requirements.txt
anthropic
apscheduler
fastapi
uvicorn[standard]
requests
beautifulsoup4
psutil
python-dotenv
